refactor: shared token manager class with the IAM related functions
