+++ ibm_cloud_sdk_core/iam_token_manager.py
@@ -15,14 +15,12 @@
 # limitations under the License.
 
 from typing import Dict, Optional
-from .jwt_token_manager import JWTTokenManager
 
-# pylint: disable=too-many-instance-attributes
-class IAMTokenManager(JWTTokenManager):
+from .iam_request_based_token_manager import IAMRequestBasedTokenManager
+
+class IAMTokenManager(IAMRequestBasedTokenManager):
     """The IAMTokenManager takes an api key and performs the necessary interactions with
     the IAM token service to obtain and store a suitable bearer token. Additionally, the IAMTokenManager
-    will retrieve bearer tokens via basic auth using a supplied client_id and client_secret pair.
-
     If the current stored bearer token has expired a new bearer token will be retrieved.
 
     Attributes:
@@ -60,13 +58,6 @@ class IAMTokenManager(JWTTokenManager):
         scope: The "scope" to use when fetching the bearer token from the IAM token server.
         This can be used to obtain an access token with a specific scope.
     """
-    DEFAULT_IAM_URL = 'https://iam.cloud.ibm.com'
-    CONTENT_TYPE = 'application/x-www-form-urlencoded'
-    OPERATION_PATH = "/identity/token"
-    REQUEST_TOKEN_GRANT_TYPE = 'urn:ibm:params:oauth:grant-type:apikey'
-    REQUEST_TOKEN_RESPONSE_TYPE = 'cloud_iam'
-    TOKEN_NAME = 'access_token'
-    SCOPE = 'scope'
 
     def __init__(self,
                  apikey: str,
@@ -78,103 +69,13 @@ class IAMTokenManager(JWTTokenManager):
                  headers: Optional[Dict[str, str]] = None,
                  proxies: Optional[Dict[str, str]] = None,
                  scope: Optional[str] = None) -> None:
-        self.apikey = apikey
-        if not url:
-            url = self.DEFAULT_IAM_URL
-        if url.endswith(self.OPERATION_PATH):
-            url = url[:-len(self.OPERATION_PATH)]
-        self.url = url
-        self.client_id = client_id
-        self.client_secret = client_secret
-        self.headers = headers
-        self.refresh_token = None
-        self.proxies = proxies
-        self.scope = scope
         super().__init__(
-            self.url, disable_ssl_verification=disable_ssl_verification, token_name=self.TOKEN_NAME)
+            url=url, client_id=client_id, client_secret=client_secret,
+            disable_ssl_verification=disable_ssl_verification, headers=headers, proxies=proxies, scope=scope)
 
-    def request_token(self) -> dict:
-        """Request an IAM OAuth token given an API Key.
+        self.apikey = apikey
 
-        If client_id and client_secret are specified use their values as a user and pass auth set
-        according to WHATWG url spec.
-
-        Returns:
-             A dictionary containing the bearer token to be subsequently used service requests.
-        """
-        headers = {
-            'Content-type': self.CONTENT_TYPE,
-            'Accept': 'application/json'
-        }
-        if self.headers is not None and isinstance(self.headers, dict):
-            headers.update(self.headers)
-
-        data = {
-            'grant_type': self.REQUEST_TOKEN_GRANT_TYPE,
-            'apikey': self.apikey,
-            'response_type': self.REQUEST_TOKEN_RESPONSE_TYPE
-        }
-
-        if self.scope is not None and self.scope:
-            data[self.SCOPE] = self.scope
-
-        auth_tuple = None
-        # If both the client_id and secret were specified by the user, then use them
-        if self.client_id and self.client_secret:
-            auth_tuple = (self.client_id, self.client_secret)
-
-        response = self._request(
-            method='POST',
-            url=(self.url + self.OPERATION_PATH) if self.url else self.url,
-            headers=headers,
-            data=data,
-            auth_tuple=auth_tuple,
-            proxies=self.proxies)
-        return response
-
-    def set_client_id_and_secret(self, client_id: str, client_secret: str) -> None:
-        """Set the client_id and client_secret.
-
-        Args:
-            client_id: The client id to be used for token requests.
-            client_secret: The client secret to be used for token requests.
-        """
-        self.client_id = client_id
-        self.client_secret = client_secret
-
-    def set_headers(self, headers: Dict[str, str]) -> None:
-        """Headers to be sent with every CP4D token request.
-
-        Args:
-            headers: Headers to be sent with every IAM token request.
-        """
-        if isinstance(headers, dict):
-            self.headers = headers
-        else:
-            raise TypeError('headers must be a dictionary')
-
-    def _save_token_info(self, token_response: dict) -> None:
-        super()._save_token_info(token_response)
-
-        self.refresh_token = token_response.get("refresh_token")
-
-    def set_proxies(self, proxies: Dict[str, str]) -> None:
-        """Sets the proxies the token manager will use to communicate with IAM on behalf of the host.
-
-        Args:
-            proxies: Proxies to use for communicating with IAM.
-            proxies.http (str, optional): The proxy endpoint to use for HTTP requests.
-            proxies.https (str, optional): The proxy endpoint to use for HTTPS requests.
-        """
-        if isinstance(proxies, dict):
-            self.proxies = proxies
-        else:
-            raise TypeError('proxies must be a dictionary')
-
-    def set_scope(self, value: str) -> None:
-        """Sets the "scope" parameter to use when fetching the bearer token from the IAM token server.
-
-        Args:
-            value: A space seperated string that makes up the scope parameter.
-        """
-        self.scope = value
+        # Set API key related data.
+        self.request_payload['grant_type'] = 'urn:ibm:params:oauth:grant-type:apikey'
+        self.request_payload['apikey'] = self.apikey
+        self.request_payload['response_type'] = 'cloud_iam'
